Gold & Transmute button QoL + Cube X button moved
QoL for the lazy and imprecise folk on PC.

Added 20% larger Transmute button for Cube and moved it closer to the inventory screen. Now it's easier to click.

Moved the X (exit) button of cube slightly to the left so now it is part of it's own UI like other X buttons.

Added 25% larger Gold button, adjusted it's positions. It is easier to click now.

Button images were upscaled via AI.