Meteor Impact Delay
Meteor impact delay updated to 1.15 seconds from 2.4 seconds.

Sound for meteor cast reduced to 0 for legacy and HD as the .flac is not open source and a replacement could not be found.

Lava rock model replaced to not exist because it causes the animation to always appear in the wrong timing.

Particles updated to only show a ring of fire for Meteor cast position and removed erroneous particles that only appeared with the wrong timing.