Splashy and Vengy
Splash Charm reworked into a radial type splash that expands from the target struck instead of a cone from the character.
- No longer double dips source damage.
Vengeance reworked so that the elemental damage is applied to the skill when the attack starts and the calc 1 - 3 fields are disabled but func35 is still used to activate the aura stats that make the elemental damage manipulation possible.
- This allows more fine control over the damage and allows it to splash properly.