Tweaking
Re did dagger on shift 4 and tweaked the other 2 around what I fleshed out there. Nova synergy to 15 from 10 and adjusted the curve slightly so it feels better when getting it and getting a few points into it. Tweak/Revert as you please I feel Dagger and Shot may be overtuned but when compared to sorc/druid/pal not really, lol.