Cleanup
Being unable to utilize D2Lint for the cube has been annoying for a while. So I fixed that too.

cubemain.txt = All uses of any that did not define the base item type (rin, amu, etc used followed by ,any) removed because it is redundent according to d2modinfo threads it can even sometimes cause issues, so just cleaned it up.

Had chatGPT write a script to enclose everything with , modifiers in quotes so that error is no longer being thrown.

Fixed the 5 broken resistances item upgrade recipes Lint found after finally clearing all of the other misc errors up.

Stacking catch all at the bottom I don't see us having a use for, it literally only catches keys and books now. Lint hates it, we don't use it, we can re-add it if we ever hit the cubes row limit, we've got a ways to go and I think we'll hit MISC or something first. Removed.

Yes this PR has gotten to be a bit massive, however, almost 95% of it is just fixing random BS and future-proofing, the changes have been condensed in the comments.