CTC & Merc Bug & Upgrade Overflows (#868)
* CTC & Merc Bug

CTC Bone Wall and Prison cause the merc to die and permanently vanish when they die or expire, cannot be rehired, cannot be resurrected. The combination of itemeffect and itemtarget don't change the fact that it still kills the merc.

Disabling them entirely as CTC skills as they have no use on the player anyways.

Fixed Poison Volley CTC requiring a corpse which is a holdover from the OG skill Poison Explosion

* Update cubemain.txt