Kings Trinity Test
Cube an ID scroll to get a charm with the Oskill

Oskill that does a physical hit and spawns 3 projectiles that circle with a few frames of delay between them converting 100% of 37% of weapon damage to an elemental damage type.

Initial hit carries all weapon effects.
Spawned projectiles do not.

Current visuals are placeholders but you can't swap to legacy to see what they are doing because of the custom icon so it's setup to show in HD.