Kings Trinity & IK Set Updates & Vengeance (#957)
* Kings Trinity Test

Cube an ID scroll to get a charm with the Oskill

Oskill that does a physical hit and spawns 3 projectiles that circle with a few frames of delay between them converting 100% of 37% of weapon damage to an elemental damage type.

Initial hit carries all weapon effects.
Spawned projectiles do not.

Current visuals are placeholders but you can't swap to legacy to see what they are doing because of the custom icon so it's setup to show in HD.

* Update missiles.txt

* Update missiles.txt

* Update skilldesc.txt

* Update Visuals

Updates the visuals for the skills to use tri-element twisters and have them combine visually for a nice gradient at the end of travel.

* kings charisma

* Update cubemain.txt

* Revert "Update cubemain.txt"

This reverts commit 1f2da56428da3bf0a558f1281a0ae9f9cdd6f9d3.

* push initial ik changes

* tweak set/items

* desc updates

* remove cube recipe

* remove kings charisma curve

* final tweaks probably

* tri-element cast overlay

add tri-elemental cast overlay for kings charisma but it could be used elsewhere if we need it for future skills/oskills

* IK weapon remove poison, add phys

* Update vengeance

* Update NHD

* SRC

* Update missiles.txt

* fix kings charisma damage and desc

* update skills to remove nhd and lower weapon carry

* Update missiles.txt

* Update skills.txt

* Update skills.txt

* Update skills.json

* Update skills.txt

* Update skills.txt

---------

Co-authored-by: Delegus <189108385+Delegus@users.noreply.github.com>
Co-authored-by: BenJz <btjezior@mtu.edu>
Co-authored-by: BenJz <contactingbenj@gmail.com>